git-svn-id: https://opengl-scut.googlecode.com/svn/trunk@455 81c3a017-244d-672b-1dc6-c8d6962b81c2

+++ Code/Scar/Scar/MySceneManager.h
@@ -29,7 +29,7 @@ public:
 	{
 		IWeapon* bullet = new BulletNode( smgr );
 		bullet->setMaterialTexture( 0, driver->getTexture( textureFileName ) );
-		bullet->SetVelocity( velocity );
+		bullet->SetVelocity( (f32)velocity );
 		bullet->SetInterval( interval );
 
 		return bullet;
